requirements-dev.txt
coverage==5.5
factory-boy==3.2.1
faker==18.3.1
pytest==6.2.2
pytest-cov==2.5.1
pytest-django==3.9.0
pytest-dotenv==0.5.2
pytest-factoryboy==2.0.3
black==23.3.0